requirements.txt
# Screaming Fixes - Broken Links & Redirect Chains Fixer

# Core
streamlit>=1.32.0
pandas>=2.0.0
python-dotenv>=1.0.0

# HTTP client (for WordPress API and Supabase)
httpx>=0.27.0
requests>=2.31.0

# LLM (optional - for AI fix suggestions on broken links)
anthropic>=0.18.0

# Analytics (optional - for usage tracking)
langsmith>=0.1.0
